pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opentools"
version = "0.1.0"
description = "an opentool registry"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "anthropic>=0.75.0",
  "google-genai>=1.56.0",
  "httpx>=0.28.1",
  "langchain>=1.2.0",
  "langchain-google-genai>=4.1.2",
  "langchain-ollama>=1.0.1",
  "langchain-openai>=1.1.6",
  "ollama>=0.6.1",
  "openai>=2.14.0",
  "pydantic>=2.12.5",
  "pydantic-ai>=1.38.0",
  "pynacl>=1.6.1",
  "python-dotenv>=1.2.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["opentools*"]
